creating schema package
