+++ packages/fluid/src/index.ts
@@ -1,6 +1,6 @@
 import { RandomCycle } from "@web-audio/patterns";
 import Synthesizer from "./instruments/synthesizer";
-import type { SynthesizerSchema, Waveform } from "./types";
+import type { DromeSchema, Waveform } from "./types";
 
 class Drome {
   private _instruments: Set<Synthesizer>;
@@ -21,11 +21,11 @@ class Drome {
     this._instruments.add(inst);
   }
 
-  getSchema() {
+  getSchema(): DromeSchema {
     return {
       instruments: Array.from(this._instruments).map((i) => i.getSchema()),
     };
   }
 }
 
-export { Drome, Synthesizer, Synthesizer as Synth, type SynthesizerSchema };
+export default Drome;
