+++ packages/fluid/src/instruments/synthesizer.ts
@@ -1,5 +1,6 @@
+import type { SynthesizerSchema } from "@web-audio/schema";
 import Instrument from "./instrument";
-import type { Drome } from "@/index";
+import type Drome from "@/index";
 import type { Waveform } from "@/types";
 
 interface SynthesizerOptions {
@@ -27,7 +28,7 @@ class Synthesizer extends Instrument {
     return this;
   }
 
-  getSchema() {
+  getSchema(): SynthesizerSchema {
     return {
       type: "synthesizer" as const,
       waveform: this._type,
