+++ packages/fluid/src/types.ts
@@ -1,4 +1,9 @@
-import Synthesizer from "./instruments/synthesizer";
+export type {
+  SynthesizerSchema,
+  DromeSchema,
+  Waveform,
+} from "@web-audio/schema";
+
 import type { scaleAliasMap } from "./utils/get-scale";
 
 type ScaleAlias = keyof typeof scaleAliasMap;
@@ -13,7 +18,4 @@ type NoteNameUpper = NaturalNote | AccidentalNote;
 type NoteName = NoteNameUpper | Lowercase<NoteNameUpper>;
 type NoteValue = `${NoteName}${number}`;
 
-type Waveform = Exclude<OscillatorType, "custom">;
-type SynthesizerSchema = ReturnType<typeof Synthesizer.prototype.getSchema>;
-
-export type { SynthesizerSchema, Waveform, ScaleAlias, NoteName, NoteValue };
+export type { ScaleAlias, NoteName, NoteValue };
