+++ packages/schema/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@web-audio/schema",
+  "type": "module",
+  "version": "0.0.0",
+  "exports": {
+    ".": "./src/index.ts"
+  }
+}
